+++ examples/test-fs-gzip-blob-storage-read.js
@@ -1,15 +1,13 @@
 #!/usr/bin/env node
 
-require("stream.pipeline-shim/auto")
+import * as stream from "node:stream"
+import * as util from "node:util"
 
-const stream = require("stream")
-const util = require("util")
-
-const {FsGzipBlobStorage} = require("../lib/fs-gzip-blob-storage")
+import {FsGzipBlobStorage} from "../lib/fs-gzip-blob-storage.js"
 
 const pipelinePromise = util.promisify(stream.pipeline)
 
-const SPOOLDIR = process.env.SPOOLDIR || "."
+const SPOOLDIR = process.env.SPOOLDIR || "spool"
 const DEBUG = Boolean(process.env.DEBUG)
 
 async function main() {