+++ examples/test-fs-gzip-blob-storage-write.js
@@ -1,15 +1,13 @@
 #!/usr/bin/env node
 
-require("stream.pipeline-shim/auto")
+import * as stream from "node:stream"
+import * as util from "node:util"
 
-const stream = require("stream")
-const util = require("util")
-
-const {FsGzipBlobStorage} = require("../lib/fs-gzip-blob-storage")
+import {FsGzipBlobStorage} from "../lib/fs-gzip-blob-storage.js"
 
 const pipelinePromise = util.promisify(stream.pipeline)
 
-const SPOOLDIR = process.env.SPOOLDIR || "."
+const SPOOLDIR = process.env.SPOOLDIR || "spool"
 const DEBUG = Boolean(process.env.DEBUG)
 
 async function main() {
@@ -18,7 +16,7 @@ async function main() {
   const key = process.argv[2]
 
   if (!key) {
-    console.error(`Usage: ${process.argv[1]} key`)
+    console.error(`Usage: ${process.argv[1]} key < file`)
     process.exit(1)
   }
 
@@ -27,8 +25,8 @@ async function main() {
 
   // extra debug trace
   if (DEBUG) {
-    for (const s of [process.stdin, writable._writable, writable]) {
-      for (const event of ["close", "data", "drain", "end", "error", "finish", "pipe", "readable", "unpipe"]) {
+    for (const s of [process.stdin, writable]) {
+      for (const event of ["close", "data", "drain", "end", "error", "finish", "pipe", "unpipe"]) {
         const name = s === process.stdin ? "stdin" : s.constructor.name
         s.on(event, arg =>
           console.debug(`${name} emitted ${event}:`, typeof arg === "object" ? arg.constructor.name : arg),