+++ test/fs-gzip-blob-storage-exclusive-errors.ts
@@ -1,12 +1,12 @@
-import zlib from "zlib"
+import zlib from "node:zlib"
 
 import {expect} from "chai"
 
-import {FsGzipBlobStorage} from "../src/fs-gzip-blob-storage"
+import {FsGzipBlobStorage} from "../src/fs-gzip-blob-storage.js"
 
-import {mockFs} from "./lib/mock-fs"
+import mockFs from "./lib/mock-fs.js"
 
-import {Before, Feature, Given, Scenario, Then, When} from "./lib/steps"
+import {Before, Feature, Given, Scenario, Then, When} from "./lib/steps.js"
 
 const STORAGEDIR = "/tmp/storage"
 
@@ -21,7 +21,7 @@ Feature("Test FsGzipBlobStorage errors for exclusive option", () => {
   Scenario("FsGzipBlobStorage tries to produce write stream when part file exists", () => {
     const testKey = "exists1"
 
-    let error: Error
+    let error: any
     let storage: FsGzipBlobStorage
 
     Before(() => {
@@ -48,7 +48,7 @@ Feature("Test FsGzipBlobStorage errors for exclusive option", () => {
   Scenario("FsGzipBlobStorage tries to produce write stream when object file exists", () => {
     const testKey = "exists2"
 
-    let error: Error
+    let error: any
     let storage: FsGzipBlobStorage
 
     Before(() => {